fix win32 build

+++ src/gfx/gfx_font_adapter_freetype2.cpp
@@ -402,10 +402,10 @@ static void decompose_ft_bitmap_mono(const FT_Bitmap& bitmap, int x, int y,
         pitch = -pitch;
     }
 
-    for (int i = 0; i < bitmap.rows; i++) {
+    for (int i = 0; i < (int)bitmap.rows; i++) {
         sl.reset_spans();
         bitset_iterator bits(buf, 0);
-        for (int j = 0; j < bitmap.width; j++) {
+        for (int j = 0; j < (int)bitmap.width; j++) {
             if (bits.bit()) sl.add_cell(x + j, cover_full)
                 ; //do nothing
             ++bits;
